Verificando se pode reutilizar a bala

+++ CloneWars/src/org/academiadecodigo/tailormoons/Bullet.java
@@ -6,11 +6,12 @@ import org.academiadecodigo.simplegraphics.graphics.Ellipse;
 import org.academiadecodigo.simplegraphics.graphics.Rectangle;
 
 
-public class Bullet{
+public class Bullet {
 
     private final int BULLET_SIZE = 10;
     private final int BULLET_SPEED = 10;
     private Ellipse bullet;
+    private boolean canReuse;
 
     public Bullet(double v, double v1, int PICTURE_MIDDLE) {
         bullet = new Ellipse(v + PICTURE_MIDDLE, v1 - 30, BULLET_SIZE, BULLET_SIZE);
@@ -20,20 +21,21 @@ public class Bullet{
 
     }
 
-    private void moving() throws InterruptedException {
-        while (bullet.getY() > 0) {
-            Thread.sleep(100);
-            bullet.fill();
-            this.move();
-        }
-        bullet.delete();
-    }
-
     public void move() {
         bullet.translate(0, -BULLET_SPEED);
+        if (bullet.getY() <= 0) {
+            bullet.delete();
+            canReuse = true;
+        }
     }
 
+    public boolean getCanReuse(){
+        return canReuse;
+    }
 
-
-
+    public void setBulletPos(int x, int y, int PICTURE_MIDDLE) {
+        this.bullet.translate(x - this.bullet.getX() + PICTURE_MIDDLE, y - this.bullet.getY());
+        this.canReuse = false;
+        this.bullet.fill();
+    }
 }
